Fix triggering and display of interactive ntoes.
 * There was some lockstep issues with the logic in the display update.
Now I use an explicit flag to state that the player is currently playing
at least one interactive instrument, so the display update needs to
refresh.
 * Only delete the cached note on the VKB when a note ends if it's the
same note. It's possible that the note that has just ended was already
overwritten by a new note before the sample playing finished, don't want
to remove the instrument from the list then, otherwise it'll never get
the release notification.