Fix reset of monitors at song change.
 * renderMonitors was relying on a local cache of the track data, but
that only gets updated when playing. If called when resetting, the
number of tracks in the cache didn’t match the new song. Instead,
don’t hold a cache of the track data, only use it when responding to
a change, and create a ‘resetMonitors’ function to use when changing
song.