Fix problem when loading in background.
Sometimes, when a song is loaded via the URL method, in a tab that
isn't the focus, when the tab comes into focus, the patterns aren't
rendered properly. This was because of the order of processing pattern
draw in relation to the font loading. It was using
requestAnimationFrame to retrigger if the font wasn't loaded, but
that's no good, as the song has started playing by then. Instead, it
now queues a callback, all of which will be called when the font is
loaded, ensuring the patterns are rendered before the view is drawn.